Move LibraryFoldersService to services package
